fix(midnight-contracts): split CLI out of wallet-info so it is browser-bundle-safe
`get-wallet-info.ts` is reached in frontend bundles via
`import("@effectstream/midnight-contracts/wallet-info")` (used by
`@effectstream/wallets` connectFromSeed). Its top-level `parseArgs`
(node:util) and `dotenv` imports poisoned the browser bundle.

Move the CLI (`main()` + the `import.meta.main` runner + its Node-only
imports) into a new `get-wallet-info-cli.ts`. `get-wallet-info.ts` now
exports only reusable library functions; `getInitialUnshieldedState` is
now exported so the CLI can consume it. No existing export removed.

Run the CLI with: bun run src/get-wallet-info-cli.ts [--seed <hex>] [--balance]
